MDEV-21117: --tc-heuristic-recover=rollback is not replication safe
(This is a pre-review commit with some cleanup as planed.

TODO: finish off cleanup and
find out by what innodb warnings are caused (suppressed in couple of tests here,
but not everywhere).

Problem:
=======
When run after master server crash --tc-heuristic-recover=rollback produces
inconsistent server state with binlog still containing transactions that were
rolled back by the option. Such way recovered server may not be used for
replication.

Fix:
===
--tc-heuristic-recover=BINLOG_TRUNCATE value to the server option is added.
It works as ROLLBACK for engines but makes its best to synchronize binlog
to truncate it when it's possible to the last committed transaction.

The BINLOG_TRUNCATE algorithm starts scanning possible a sequences of binlog files
starting from the binlog checkpoint file to find a truncate position candidate.
The candidate corresponds to the beginning of the first GTID event whose
transaction is not found to have been committed. There must be no committed transactions
beyound this position *but* non-transactional events are possible. This is also
checked to warn when the case.
Rollback or completion of partial commit (in the multi-engine) case is done regardless,
as well as the binlog GTID state is adjusted accordingly.
Also a STOP binlog event is written at the end of truncated binary log.